PHP 8.1 | WPSEO_Sitemap_Cache_Data: implement magic serialization methods
PHP 7.4 introduced the magic `__serialize()` and `__unserialize()` methods as an improved way to handle object serialization compared to the `Serializable` interface and/or the magic `__sleep()` and __wakeup()` methods.

As of PHP 8.1, implementing the `Serializable` interface is deprecated, unless the class also implements the new magic methods.

The `WPSEO_Sitemap_Cache_Data` class implements the `Serializable` interface and would throw the following deprecation notice on PHP 8.1:
```
Deprecated: WPSEO_Sitemap_Cache_Data implements the Serializable interface, which is deprecated. Implement __serialize() and __unserialize() instead (or in addition, if support for old PHP versions is necessary)
```

As the minimum supported PHP version is below 7.4, this means that - for the time being - both the `Serializable` interface as well as the magic methods need to be implemented.

For PHP 7.4 and higher, the magic methods will take precedence over the `Serializable` implementation and the `serialize()` and `unserialize()` methods will no longer be used.

Once PHP 7.4 becomes the minimum supported PHP version, these methods should be removed.

Note: the `Serializable` interface will be removed from PHP in PHP 9.0 and, while the interface can be polyfilled, such a polyfill will have **no effect** on the serialization of the object.

The existing `WPSEO_Sitemaps_Cache_Data_Test::test_serialize_unserialize()` integration test covers this change.

Includes a small adjustment to the `WPSEO_Sitemaps_Cache_Test::test_transient_cache_data_object()` to allow for WP handling the two types of serialization differently.

Includes fixing up the parameter name for the existing `unserialize()` method to be in line with the documented parameter name in PHP (for named parameter support in PHP 8.0).

Includes fixing the `@covers` tags for the `WPSEO_Sitemaps_Cache_Data_Test::test_serialize_unserialize()` test.

Refs:
* https://wiki.php.net/rfc/custom_object_serialization
* https://wiki.php.net/rfc/phase_out_serializable
* https://www.php.net/language.oop5.magic#object.serialize
* https://www.php.net/manual/en/class.serializable.php
* https://www.phpinternalsbook.com/php5/classes_objects/serialization.html