WPSEO_Sitemaps_Renderer: fix escaping of URLs in the sitemap
As things were, the URLs which were being added to the sitemap would be incorrectly escaped.

* The `loc` URL would be escaped before the encoding was checked, which could lead to `&amp&amp;` type of artifacts in the URL in the sitemap.
* The image `src` URLs would be escaped with `esc_html()` instead of `esc_url()`.
* In both cases, the schema (protocol) for schema-relative URLs would not be added, which the sitemap specs explicitly state a protocol is required.

This introduces a new `WPSEO_Sitemaps_Renderer::encode_and_escape()` method which handles the encoding and escaping in the right order and is subsequently used in the `WPSEO_Sitemaps_Renderer::get_sitemap()` method.

The previously added tests for the URL "encoding" already largely cover this change.

For the additional changes to the `get_sitemap()` method, the `loc` URL in that test has been adjusted and a few additional image URLs have been added. Together these changes  will safeguard that the encoding and escaping is done correctly.