Don't mock the SUT
Once we set a list of `redefinable-internals` in the patchwork.json
the WebPage test-cases broke, as patchwork now tries to proxy the
`add_image` function call because the instance is a partial mock.
This fails because the function expects a reference but patchwork
can only provide a value. We didn't use any of the mock features here
so I removed the mock and created a real instance instead.