PHP 8.1 | Tests: temporarily ignore select test failures [3]
The test failures being handled by setting an expectation for a PHP deprecation notice are on the list to be fixed via the "Filter extension"-CS action list.

Once the particular issue affecting these tests has been fixed, these expectations can be removed (and this is safeguarded by the expectation as the test will start to fail because the deprecation will no longer occur).

Note: the call to the `_deprecated_function()`  function isn't actually part of the test - we're allowing for it to be called, not _expecting_ it to always be called, I've updated the test to _stub_ the function instead setting an expectation for it.
This is necessary, for now, for the tests to pass, but at the same time, it is the better way to handle this anyway, so only the conditional "expectDeprecation" block should be removed when the PHP 8.1 issue is fixed, not this change.