GH Actions/JS: split the workflow into two + separate test builds per package
The JS workflow was the slowest workflow, with a runtime of easily 10 minutes.

This commit fixes that by making the runs more modular.

* The JS lint task will now be run in a separate workflow. No other changes were made to the lint job.
* The JS test task has now been split into two different jobs:
    1. One job for small packages.
    2. One job for "big" packages.

The "small" packages job will:
* Run on select push events, PRs and on demand.
* NOT run via the cron job.
* Run the full test suite for each package, always.
* Run the tests for each package via a matrix, resulting in a separate build for each package.
* Runs with `fail-fast: false` as these runs are all unique so we don't want to job to stop on the first failure.

The "big" packages job will:
* Run on select push events, PRs and on demand AND via a nightly cron job.
* Will run a partial test suite for "normal" (push/PR) events, only running the relevant tests based on the changed files between the base branch and the current branch.
* Will run the full test suite for each package once a day via the cron job.
* Will run the tests for each package via a matrix, resulting in a separate build for each package.
    For now, there is only one package in this matrix, but the workflow has been set up to allow for moving packages between the two jobs in a straight-forward manner.
* Runs with `fail-fast: false` as these runs are all unique so we don't want to job to stop on the first failure.
* Only download the premium config for the `yoastseo` package, which actually needs it.

Note: in contrast to information found _"on the web"_ (and in contrast to the PHP CS workflow), the "base branch" for the changed file determination needs to exist in the CI local git repo for the `--changedSince` setting to work.
Using `origin/trunk` does not work (confirmed via testing the workflow).

This means the base branch needs to be checked out after having been fetched. After which we return back to the _real branch_ under test with a second checkout.

End result: the run time of these workflows is now somewhere between 2-3 minutes.