Meta_Surface::for_url(): fix undefined array index warnings and more
When `[wp_]parse_url()` is used on a seriously malformed URL or an invalid input, the output may not be an array and if it is an array, it may not contain all the expected array keys.

This adds a safeguard to this method to prevent PHP notices in case the passed URL parameter falls into that category.

After this, the tests we added in the first commit, will now pass.

Co-authored-by: Maarten Leenders <26220788+maartenleenders@users.noreply.github.com>