GH Actions/lint: fix failing builds on PHP < 7.2
[Composer 2.3](https://blog.packagist.com/composer-2-3/) was just released and has a minimum PHP requirement of PHP 7.2.

In effect that means that the "efficiency tweak" to not re-install the dependencies after switching the PHP version in this script, no longer works as the Composer `autoload` files generated with Composer 2.3 on PHP 7.2 will contain PHP syntax which is incompatible with PHP < 7.2.

The reason for not doing a new `composer install` in this script after switching PHP versions was that it isn't actually needed, as the only dependency we need from that point forward is the PHP Parallel Lint dependency and that is compatible with PHP 5.3 - current.

To fix this, there are three choices:
1. Re-run `composer install` (do a full new install) after switching to the PHP version on which we want to do the linting.
    This may run into trouble with some dependencies not locked at PHP 5.6 compatible versions.
2. Run `composer du` (regenerate just the autoload files) after switching to the PHP version on which we want to do the linting.
    This should work fine.
3. Force `setup-php` to install Composer 2.2 during the first PHP setup so the autoload file will be cross-version compatible.
    This should also work fine.

As the efficiency tweak of preventing a full second `composer install` is still valid, I've chosen to implement option 2.

While option 3 would also be perfectly valid, it would introduce an inconsistency when comparing this script to other scripts used and always raise questions, so the second option is the one which is more self-documenting.