Robots_Txt_Helper: use OS-based new lines
As things were, the `Robots_Txt_Helper` was mixing the use of `"\n"` (Linux new lines) and `PHP_EOL` (OS-based new lines).

This could (on non-Linux machines) result in the text string being generated for `robots.txt` having mixed line endings, which is never a good idea.

The [RFC regarding `robots.txt`](https://datatracker.ietf.org/doc/rfc9309/) specifies the _use_ of `EOL` characters in specific places, but doesn't specify whether those should be OS-based or `\n`.

As the `robots.txt` file is, in rare cases, also being called up by humans, using `PHP_EOL` or `"\r\n"` should probably be the preferred choice of line ending to keep the file human readable.

With this in mind, this PR proposes to standardize on `PHP_EOL`, which was already used in a number of places.

Note: the OS on which `PHP_EOL` is based will be the OS of the machine on which the code is being run, i.e. the machine _serving_ the "page", not the OS of the user potentially _reading_ the file.

Includes updating the text expectations to use `PHP_EOL` as well (and making those test expectations a little more human readable as well when needed).