WebPage_Test: remove stray property setting [1]
The `website_hash` property is set on the `ID_Helper` class at the end of the `WebPage_Test::set_up()` method.

However, in the whole of the YoastSEO code base, no `$website_hash` property exists, nor any references to a `->website_hash` property, only to a class constant by the same name.
.

I suspect this is a "stray"/copy-paste test-only bug and that the property assignment should just be removed.