Composer/Tests: update test dependencies
This updates the minimum version of the WP Test Utils dependency and pulls in the latest PHP 7.2 compatible versions of all test dependencies.

Note: PHPUnit Polyfills 1.1.0 includes a new polyfill for the `assertObject[Not]HasProperty()` methods.
The assertions which these new assertions replace are not used in this test suite, so no test changes are needed.