CI: Build all packages before running tests
Packages might depend on other packages. Because those packages now have their
`main` property pointing to the build version, those build versions need to
exist to make the tests pass.

This doesn't build the plugin js itself. It has dependencies on javascript from
 WordProof which would have be built with composer prefixing. That would only
lengthen the builds. But the plugin JavaScript isn't needed, because it is the
 top of the foodchain and will never be a dependency for any of the other packages.