Make stories on both the docs and their individual story pages interactive
By replacing the DocsPage in every component story we circumvent the default Stories component from passing `__forceInitialArgs` (true) to each Story.
This also enables us to skip the Factory there, that is a duplicate of the Primary story, which was not there in Storybook 6.
The downside being that we have to pass each story manually.

Some stories required some additional fixes to become interactive in both the docs page and their individual story page. It is important to:
* use `updateArgs` whenever an argument can change from an interaction
* not mix `updateArgs` with `setState` at all, instead you can add some "hidden" args