Expose `useToastContext` to allow children to dismiss toast
Having a button inside the toast which dismisses itself, requires a lot of
additional boilerplate. Exposing the predefined `handleDismiss` takes
this concern away from anyone implementing the toast.

Next to that, the Close button has a 150ms delay before actually closing,
which consumers of the Toast would have to replicate for consistent
behaviour. This ensures a consistent delay everywhere if we decide to
adjust such delay at a later stage.